Added 'API Reference' section (#17)
Made a separate docs page for the APIs and to explain them in depth.

https://buildernet-website-git-docs-api-flashbots.vercel.app/docs/api

This provides clarity and is particularly useful for orderflow providers
and users.

Some information is now becoming a bit redundant (i.e. in "how to send
orderflow" and "orderflow encryption and sharing"), and these sections
can get cleaned up in follow-up PRs.