Fix broken links and standardize links (#28)
This PR fixes a broken link on the Welcome to BuilderNet page.

The issue was caused by a relative link that incorrectly pointed to
`https://buildernet.org/how-to-participate` instead of the correct path:
`https://buildernet.org/docs/how-to-participate`.


![image](https://github.com/user-attachments/assets/de40516b-3169-4cf0-97cf-cb3c700626ea)

## How to reproduce the bug 

1. Navigate to the _docs_ section via topbar, or select _Documentation_
or _What is BuilderNet?_ from the main page.


![image](https://github.com/user-attachments/assets/ab899859-75f6-4118-86bb-819407890f19)


![image](https://github.com/user-attachments/assets/5514df6f-0251-4da7-855e-80cfbe4dddb7)

2.	Click on the `contribute` link.


![image](https://github.com/user-attachments/assets/23d95e2b-1f0a-41ec-b256-8c15100349dc)

3. This will lead to a “Page Not Found” error, as the link incorrectly
points to `https://buildernet.org/how-to-participate` instead of the
correct `https://buildernet.org/docs/how-to-participate`.


![image](https://github.com/user-attachments/assets/1f3bafb8-e800-48f4-a14f-25c334394451)

(Tested both on Chrome/Mozilla) 

## Fix 

- Replaced the broken relative link with the correct “absolute” link:
`[contribute](/docs/how-to-participate)`
- Updated other links on the page to follow the standard format,
ensuring they do not include a trailing slash.

## Test

Tested locally via 

```bash
make docker
```

Then access the page at `localhost:3000/docs`